requirements/requirements_dev.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-countries==6.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
djangorestframework==3.11.1
idna==2.10
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
